+++ R/progress.R
@@ -1,9 +1,6 @@
 ##' @importFrom httpcache uncached
 ##' @importFrom utils txtProgressBar setTxtProgressBar
-pollProgress <- function (progress_url, wait=1) {
-    ## TODO: Take Location url, and on.exit(print(it)) so that user can interrupt
-    ## progress polling
-
+pollProgress <- function (progress_url, wait=.5) {
     ## Configure polling interval. Will increase by rate (>1) until reaches max
     max.wait <- 30
     increase.by <- 1.2
@@ -18,14 +15,16 @@ pollProgress <- function (progress_url, wait=1) {
 
     status <- uncached(as.numeric(crGET(progress_url)$progress))
     setTxtProgressBar(pb, status)
-    while (status < 100 && timer(starttime) < timeout) {
+    while (status >= 0 && status < 100 && timer(starttime) < timeout) {
         Sys.sleep(wait)
         status <- uncached(as.numeric(crGET(progress_url)$progress))
         setTxtProgressBar(pb, status)
         wait <- min(max.wait, wait * increase.by)
     }
 
-    if (status != 100) {
+    if (status < 0) {
+        halt("There was an error on the server. Please contact support@crunch.io")
+    } else if (status != 100) {
         halt('Your process is still running on the server. It is currently ',
             round(status), '% complete. Check `httpcache::uncached(crGET("',
             progress_url, '"))` until it reports 100% complete')